+++ test/LionWeb.Core.Test/Notification/Replicator/Annotation/TwowayTests.cs
@@ -17,7 +17,7 @@
 
 namespace LionWeb.Core.Test.Notification.Replicator.Annotation;
 
-using Languages.Generated.V2024_1.Shapes.M2;
+using Languages.Generated.V2024_1.TestLanguage;
 
 [TestClass]
 public class TwowayTests : TwowayReplicatorTestsBase
@@ -27,14 +27,14 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationAdded_Multiple_Only()
     {
-        var node = new Geometry("a");
+        var node = new TestPartition("a");
 
-        var clone = new Geometry("a");
+        var clone = new TestPartition("a");
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
 
-        var added = new BillOfMaterials("added");
+        var added = new TestAnnotation("added");
         node.AddAnnotations([added]);
 
         AssertEquals([node], [clone]);
@@ -44,16 +44,16 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationAdded_Multiple_First()
     {
-        var node = new Geometry("a");
-        node.AddAnnotations([new BillOfMaterials("bof")]);
+        var node = new TestPartition("a");
+        node.AddAnnotations([new TestAnnotation("bof")]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("bof")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("bof")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
 
-        var added = new BillOfMaterials("added");
+        var added = new TestAnnotation("added");
         node.InsertAnnotations(0, [added]);
 
         AssertEquals([node], [clone]);
@@ -63,16 +63,16 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationAdded_Multiple_Last()
     {
-        var node = new Geometry("a");
-        node.AddAnnotations([new BillOfMaterials("bof")]);
+        var node = new TestPartition("a");
+        node.AddAnnotations([new TestAnnotation("bof")]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("bof")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("bof")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
 
-        var added = new BillOfMaterials("added");
+        var added = new TestAnnotation("added");
         node.InsertAnnotations(1, [added]);
 
         AssertEquals([node], [clone]);
@@ -80,23 +80,24 @@ public class TwowayTests : TwowayReplicatorTestsBase
     }
 
     [TestMethod]
+    [Ignore("Not possible with TestLanguage")]
     public void AnnotationAdded_Deep()
     {
-        var node = new Geometry("a");
+        var node = new TestPartition("a");
 
-        var clone = new Geometry("a");
+        var clone = new TestPartition("a");
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
 
-        var added = new BillOfMaterials("added")
-        {
-            AltGroups = [new MaterialGroup("mg") { MatterState = MatterState.gas }]
-        };
-        node.AddAnnotations([added]);
+        // var added = new TestAnnotation("added")
+        // {
+        //     AltGroups = [new MaterialGroup("mg") { MatterState = MatterState.gas }]
+        // };
+        // node.AddAnnotations([added]);
 
         AssertEquals([node], [clone]);
-        Assert.AreNotSame(added, clone.GetAnnotations()[0]);
+        // Assert.AreNotSame(added, clone.GetAnnotations()[0]);
     }
 
     #endregion
@@ -106,12 +107,12 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationDeleted_Multiple_Only()
     {
-        var deleted = new BillOfMaterials("deleted");
-        var node = new Geometry("a");
+        var deleted = new TestAnnotation("deleted");
+        var node = new TestPartition("a");
         node.AddAnnotations([deleted]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("deleted")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("deleted")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
@@ -124,12 +125,12 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationDeleted_Multiple_First()
     {
-        var deleted = new BillOfMaterials("deleted");
-        var node = new Geometry("a");
-        node.AddAnnotations([deleted, new BillOfMaterials("bof")]);
+        var deleted = new TestAnnotation("deleted");
+        var node = new TestPartition("a");
+        node.AddAnnotations([deleted, new TestAnnotation("bof")]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("deleted"), new BillOfMaterials("bof")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("deleted"), new TestAnnotation("bof")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
@@ -142,12 +143,12 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationDeleted_Multiple_Last()
     {
-        var deleted = new BillOfMaterials("deleted");
-        var node = new Geometry("a");
-        node.AddAnnotations([new BillOfMaterials("bof"), deleted]);
+        var deleted = new TestAnnotation("deleted");
+        var node = new TestPartition("a");
+        node.AddAnnotations([new TestAnnotation("bof"), deleted]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("bof"), new BillOfMaterials("deleted")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("bof"), new TestAnnotation("deleted")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
@@ -164,14 +165,14 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationMovedFromOtherParent_Multiple()
     {
-        var moved = new BillOfMaterials("moved");
-        var origin = new CompositeShape("origin");
+        var moved = new TestAnnotation("moved");
+        var origin = new LinkTestConcept("origin");
         origin.AddAnnotations([moved]);
-        var node = new Geometry("a") { Shapes = [origin] };
+        var node = new TestPartition("a") { Links = [origin] };
 
-        var cloneOrigin = new CompositeShape("origin");
-        cloneOrigin.AddAnnotations([new BillOfMaterials("moved")]);
-        var clone = new Geometry("a") { Shapes = [cloneOrigin] };
+        var cloneOrigin = new LinkTestConcept("origin");
+        cloneOrigin.AddAnnotations([new TestAnnotation("moved")]);
+        var clone = new TestPartition("a") { Links = [cloneOrigin] };
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
@@ -188,12 +189,12 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationMovedInSameParent_Forward()
     {
-        var moved = new BillOfMaterials("moved");
-        var node = new Geometry("a");
-        node.AddAnnotations([moved, new BillOfMaterials("bof")]);
+        var moved = new TestAnnotation("moved");
+        var node = new TestPartition("a");
+        node.AddAnnotations([moved, new TestAnnotation("bof")]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("moved"), new BillOfMaterials("bof")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("moved"), new TestAnnotation("bof")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);
 
@@ -206,12 +207,12 @@ public class TwowayTests : TwowayReplicatorTestsBase
     [TestMethod]
     public void AnnotationMovedInSameParent_Backward()
     {
-        var moved = new BillOfMaterials("moved");
-        var node = new Geometry("a");
-        node.AddAnnotations([new BillOfMaterials("bof"), moved]);
+        var moved = new TestAnnotation("moved");
+        var node = new TestPartition("a");
+        node.AddAnnotations([new TestAnnotation("bof"), moved]);
 
-        var clone = new Geometry("a");
-        clone.AddAnnotations([new BillOfMaterials("bof"), new BillOfMaterials("moved")]);
+        var clone = new TestPartition("a");
+        clone.AddAnnotations([new TestAnnotation("bof"), new TestAnnotation("moved")]);
 
         var (replicator, cloneReplicator) = CreateReplicators(node, clone);