+++ test/LionWeb.Core.Test/Notification/Replicator/Reference/AddedTests.cs
@@ -17,7 +17,7 @@
 
 namespace LionWeb.Core.Test.Notification.Replicator.Reference;
 
-using Languages.Generated.V2025_1.Shapes.M2;
+using Languages.Generated.V2024_1.TestLanguage;
 
 [TestClass]
 public class AddedTests : ReplicatorTestsBase
@@ -25,15 +25,15 @@ public class AddedTests : ReplicatorTestsBase
     [TestMethod]
     public void Multiple_Only()
     {
-        var bof = new BillOfMaterials("bof");
-        var line = new Line("line");
-        var originalPartition = new Geometry("a") { Shapes = [line] };
-        originalPartition.AddAnnotations([bof]);
+        var bof = new LinkTestConcept("bof");
+        var line = new LinkTestConcept("line");
+        var originalPartition = new TestPartition("a") { Links =  [line] };
+        originalPartition.AddLinks([bof]);
 
         var clonedPartition = ClonePartition(originalPartition);
         CreatePartitionReplicator(clonedPartition, originalPartition);
 
-        bof.AddMaterials([line]);
+        bof.AddReference_0_n([line]);
 
         AssertEquals([originalPartition], [clonedPartition]);
     }
@@ -41,16 +41,16 @@ public class AddedTests : ReplicatorTestsBase
     [TestMethod]
     public void Multiple_First()
     {
-        var circle = new Circle("circle");
-        var bof = new BillOfMaterials("bof") { Materials = [circle] };
-        var line = new Line("line");
-        var originalPartition = new Geometry("a") { Shapes = [line, circle] };
-        originalPartition.AddAnnotations([bof]);
+        var circle = new LinkTestConcept("circle");
+        var bof = new LinkTestConcept("bof") { Reference_0_n =  [circle] };
+        var line = new LinkTestConcept("line");
+        var originalPartition = new TestPartition("a") { Links =  [line, circle] };
+        originalPartition.AddLinks([bof]);
 
         var clonedPartition = ClonePartition(originalPartition);
         CreatePartitionReplicator(clonedPartition, originalPartition);
 
-        bof.InsertMaterials(0, [line]);
+        bof.InsertReference_0_n(0, [line]);
 
         AssertEquals([originalPartition], [clonedPartition]);
     }
@@ -58,16 +58,16 @@ public class AddedTests : ReplicatorTestsBase
     [TestMethod]
     public void Multiple_Last()
     {
-        var circle = new Circle("circle");
-        var bof = new BillOfMaterials("bof") { Materials = [circle] };
-        var line = new Line("line");
-        var originalPartition = new Geometry("a") { Shapes = [line, circle] };
-        originalPartition.AddAnnotations([bof]);
+        var circle = new LinkTestConcept("circle");
+        var bof = new LinkTestConcept("bof") { Reference_0_n =  [circle] };
+        var line = new LinkTestConcept("line");
+        var originalPartition = new TestPartition("a") { Links =  [line, circle] };
+        originalPartition.AddLinks([bof]);
 
         var clonedPartition = ClonePartition(originalPartition);
         CreatePartitionReplicator(clonedPartition, originalPartition);
 
-        bof.InsertMaterials(1, [line]);
+        bof.InsertReference_0_n(1, [line]);
 
         AssertEquals([originalPartition], [clonedPartition]);
     }
@@ -75,19 +75,19 @@ public class AddedTests : ReplicatorTestsBase
     [TestMethod]
     public void Single()
     {
-        var circle = new Circle("circle");
-        var od = new OffsetDuplicate("od");
-        var originalPartition = new Geometry("a") { Shapes = [od, circle] };
+        var circle = new LinkTestConcept("circle");
+        var od = new LinkTestConcept("od");
+        var originalPartition = new TestPartition("a") { Links =  [od, circle] };
 
         var clonedPartition = ClonePartition(originalPartition);
         CreatePartitionReplicator(clonedPartition, originalPartition);
 
-        od.Source = circle;
+        od.Reference_0_1 = circle;
 
         AssertEquals([originalPartition], [clonedPartition]);
 
-        var clonedOffsetDuplicate = (OffsetDuplicate)clonedPartition.Shapes[0];
-        var clonedCircle = (Circle)clonedPartition.Shapes[1];
-        Assert.AreSame(clonedCircle, clonedOffsetDuplicate.Source);
+        var clonedOffsetDuplicate = clonedPartition.Links[0];
+        var clonedCircle = clonedPartition.Links[1];
+        Assert.AreSame(clonedCircle, clonedOffsetDuplicate.Reference_0_1);
     }
 }